+++ cli/lib/exec/run.ts
@@ -71,6 +71,10 @@ const processRunOptions = (options: any = {}): string[] => {
     args.push('--env', options.env)
   }
 
+  if (options.expose) {
+    args.push('--expose', options.expose)
+  }
+
   if (options.exit === false) {
     args.push('--no-exit')
   }