+++ packages/driver/src/cy/commands/origin/index.ts
@@ -5,7 +5,7 @@ import { Validator } from './validator'
 import { isFunction } from 'lodash'
 import { createUnserializableSubjectProxy } from './unserializable_subject_proxy'
 import { serializeRunnable } from './util'
-import { preprocessConfig, preprocessEnv, syncConfigToCurrentOrigin, syncEnvToCurrentOrigin } from '../../../util/config'
+import { preprocessConfig, preprocessEnv, preprocessExpose, syncConfigToCurrentOrigin, syncEnvToCurrentOrigin, syncExposeToCurrentOrigin } from '../../../util/config'
 import { $Location } from '../../../cypress/location'
 import { LogUtils } from '../../../cypress/log'
 import logGroup from '../../logGroup'
@@ -129,8 +129,9 @@ export default (Commands, Cypress: InternalCypress.Cypress, cy: Cypress.cy, stat
           _resolve({ subject, unserializableSubjectType })
         }
 
-        const onSyncGlobals = ({ config, env }) => {
+        const onSyncGlobals = ({ config, env, expose }) => {
           syncConfigToCurrentOrigin(config)
+          syncExposeToCurrentOrigin(expose)
           if (Cypress.config('allowCypressEnv')) {
             syncEnvToCurrentOrigin(env)
           }
@@ -187,6 +188,7 @@ export default (Commands, Cypress: InternalCypress.Cypress, cy: Cypress.cy, stat
             communicator.toSpecBridge(origin, 'initialize:cypress', {
               config: preprocessConfig(Cypress.config()),
               env: Cypress.config('allowCypressEnv') ? preprocessEnv(Cypress.env()) : undefined,
+              expose: preprocessExpose(Cypress.expose()),
               isProtocolEnabled: Cypress.state('isProtocolEnabled'),
             })
 
@@ -232,6 +234,7 @@ export default (Commands, Cypress: InternalCypress.Cypress, cy: Cypress.cy, stat
                 },
                 config: preprocessConfig(Cypress.config()),
                 env: Cypress.config('allowCypressEnv') ? preprocessEnv(Cypress.env()) : undefined,
+                expose: preprocessExpose(Cypress.expose()),
                 logCounter: LogUtils.getCounter(),
               })
             } catch (err: any) {